refactor: cleanup FE payments folder
