+++ frontend/src/features/public-form/components/FormPaymentPage/components/index.tsx
@@ -1,0 +1,4 @@
+export * from './CreatePaymentIntentFailureBlock'
+export * from './PaymentItemDetailsBlock'
+export * from './PaymentStack'
+export * from './PaymentSuccessSvgr'
