+++ frontend/src/features/public-form/components/FormPaymentPage/stripe/components/index.tsx
@@ -1,4 +1,3 @@
 export * from './DownloadReceiptBlock'
 export * from './GenericMessageBlock'
 export * from './PaymentBlock'
-export * from './PaymentStack'
